cmapgeo

This package is a repository of geodata files representing various geographic and administrative boundaries in the Chicago region, maintained by the Chicago Metropolitan Agency for Planning (CMAP). Data is in the sf (simple features) format.

Details

Detailed documentation can be viewed at https://cmap-repos.github.io/cmapgeo.

Please report issues and suggest improvements at https://github.com/CMAP-REPOS/cmapgeo/issues.

See also

Author

Maintainer: Daniel Comeaux dcomeaux@cmap.illinois.gov (ORCID)

Authors:

Noel Peterson (ORCID)

Other contributors:

Chicago Metropolitan Agency for Planning [copyright holder, funder]