County FIPS Code Lists

A named list comprising four named vectors, each of which contains the FIPS codes (as 5-character strings) for the counties in one of the four primary geographic divisions of CMAP's travel model and socioeconomic forecasts: the 7-county CMAP region, as well as the external counties in Illinois, Indiana, and Wisconsin.

county_fips_codes

Format

An object of class list of length 5.

Examples

# 7-county CMAP region
county_fips_codes$cmap

#>    Cook  DuPage    Kane Kendall    Lake McHenry    Will 
#> "17031" "17043" "17089" "17093" "17097" "17111" "17197" 

# Travel model non-CMAP Illinois counties
county_fips_codes$xil

#>     Boone    DeKalb    Grundy  Kankakee   LaSalle       Lee      Ogle Winnebago 
#>   "17007"   "17037"   "17063"   "17091"   "17099"   "17103"   "17141"   "17201" 

# Travel model Indiana counties
county_fips_codes$xin

#>    Lake LaPorte  Porter 
#> "18089" "18091" "18127" 

# Travel model Wisconsin counties
county_fips_codes$xwi

#>  Kenosha   Racine Walworth 
#>  "55059"  "55101"  "55127" 

# Chicago-Naperville-Elgin, IL-IN-WI MSA counties
county_fips_codes$msa

#>    Cook_IL  DeKalb_IL  DuPage_IL  Grundy_IL    Kane_IL Kendall_IL    Lake_IL 
#>    "17031"    "17037"    "17043"    "17063"    "17089"    "17093"    "17097" 
#> McHenry_IL    Will_IL  Jasper_IN    Lake_IN  Newton_IN  Porter_IN Kenosha_WI 
#>    "17111"    "17197"    "18073"    "18089"    "18111"    "18127"    "55059" 

# Get the names of the CMAP counties
names(county_fips_codes$cmap)

#> [1] "Cook"    "DuPage"  "Kane"    "Kendall" "Lake"    "McHenry" "Will"   

# Get all travel model counties in a single vector
unlist(county_fips_codes[c("cmap", "xil", "xin", "xwi")])

#>     cmap.Cook   cmap.DuPage     cmap.Kane  cmap.Kendall     cmap.Lake 
#>       "17031"       "17043"       "17089"       "17093"       "17097" 
#>  cmap.McHenry     cmap.Will     xil.Boone    xil.DeKalb    xil.Grundy 
#>       "17111"       "17197"       "17007"       "17037"       "17063" 
#>  xil.Kankakee   xil.LaSalle       xil.Lee      xil.Ogle xil.Winnebago 
#>       "17091"       "17099"       "17103"       "17141"       "17201" 
#>      xin.Lake   xin.LaPorte    xin.Porter   xwi.Kenosha    xwi.Racine 
#>       "18089"       "18091"       "18127"       "55059"       "55101" 
#>  xwi.Walworth 
#>       "55127"