Tract-to-CCA Crosswalk

This table contains a set of factors to apportion Census tract-level data among Chicago Community Areas (CCAs). Seperate factors are provided for apportioning housing unit, household, and population attributes. All factors were determined by calculating the percentage of a tract's housing units, households and population that were located in each of its component blocks, according to the 2010 Decennial Census, and then assigning each block to a CCA (based on the location of the block's centroid point).

xwalk_tract2cca

Format

A tibble with 805 rows and 5 variables:

geoid_tract

Unique 11-digit tract ID, assigned by the Census Bureau. Corresponds to tract_sf. Character.

cca_num

Numeric CCA ID, as assigned by the City of Chicago. Corresponds to cca_sf. Integer.

hu_pct

Proportion of the tract's housing units (occupied or vacant) located in the specified CCA. Multiply this by a tract-level measure of a housing attribute (e.g. vacant homes) to estimate the CCA's portion. Double.

hh_pct

Proportion of the tract's households (i.e. occupied housing units) living in the specified CCA. Multiply this by a tract-level measure of a household attribute (e.g. car-free households) to estimate the CCA's portion. Double.

pop_pct

Proportion of the tract's total population (including group quarters) living in the specified CCA. Multiply this by a tract-level measure of a population attribute (e.g. race/ethnicity) to estimate the CCA's portion. Double.

Details

Generally speaking, tract boundaries align neatly with CCA boundaries as they tend to follow similar features (e.g. rivers, major roads, rail lines) but there are cases where the population, households and/or housing units in a tract are split across multiple CCAs, or else are partially within the City of Chicago and partially outside of it. For that reason, it is not appropriate to use a one-to-one tract-to-CCA assignment to apportion Census data among CCAs, and this crosswalk (or xwalk_blockgroup2cca for block group-level data) should be used instead.

To use this crosswalk effectively, Census data should be joined to it (not vice versa, since tract IDs appear multiple times in this table). Once the data is joined, it should be multiplied by the appropriate factor (depending whether the data of interest is measured at the housing unit, household or person level), and then the result should be summed by CCA. If calculating rates, this should only be done after the counts have been summed to CCA. The resulting table can then be joined to cca_sf for mapping, if desired.

If your data is also available at the block group level, it is recommended that you use that with xwalk_blockgroup2cca instead of the tract-level allocation.

Examples

library(tidyverse)

# View the tracts with population not fully contained in a single CCA
filter(xwalk_tract2cca, pop_pct < 1)

#> # A tibble: 22 x 5
#>    geoid_tract cca_num  hu_pct  hh_pct pop_pct
#>    <chr>         <int>   <dbl>   <dbl>   <dbl>
#>  1 17031090200       9 0.983   0.984   0.987  
#>  2 17031090300       9 0.958   0.956   0.951  
#>  3 17031520500      52 0.997   0.997   0.996  
#>  4 17031520500      55 0.00261 0.00270 0.00395
#>  5 17031520600      52 0.993   0.992   0.992  
#>  6 17031520600      55 0.00726 0.00760 0.00786
#>  7 17031570500      57 0.992   0.992   0.988  
#>  8 17031570500      62 0.00775 0.00844 0.0122 
#>  9 17031770902      76 0.0808  0.0768  0.0557 
#> 10 17031810400      10 0.0723  0.0703  0.0525 
#> # … with 12 more rows

# Estimate CCA-level transit mode share from tract-level ACS data
df_tract <- tidycensus::get_acs(
  "tract", state = "IL", county = "031", table = "B08006",
  year = 2019, survey = "acs5", output = "wide"
) %>%
  rename(workers = B08006_001E, transit = B08006_008E) %>%
  select(GEOID, workers, transit)

#> Getting data from the 2015-2019 5-year ACS

#> Loading ACS5 variables for 2019 from table B08006. To cache this dataset for faster access to ACS tables in the future, run this function with `cache_table = TRUE`. You only need to do this once per ACS dataset.

#> No encoding supplied: defaulting to UTF-8.

#> Error: You have supplied an invalid or inactive API key. To obtain a valid API key, visit https://api.census.gov/data/key_signup.html. To activate your key, be sure to click the link provided to you in the email from the Census Bureau that contained your key.

df_cca <- xwalk_tract2cca %>%
  left_join(df_tract, by = c("geoid_tract" = "GEOID")) %>%
  mutate(transit = transit * pop_pct,
         workers = workers * pop_pct) %>%
  group_by(cca_num) %>%
  summarize_at(vars(transit, workers), sum) %>%
  mutate(transit_commute_pct = transit / workers)

#> Error in is.data.frame(y): object 'df_tract' not found

df_cca

#> Error in eval(expr, envir, enclos): object 'df_cca' not found

# Join to cca_sf for mapping
cca_sf %>%
  left_join(df_cca, by = "cca_num") %>%
  ggplot() +
    geom_sf(aes(fill = transit_commute_pct)) +
    scale_fill_viridis_c() +
    theme_void()

#> Error in check_join(x, y): object 'df_cca' not found